#2 mod: change from mysql to mysql2
