+++ controllers/search/searchController.js
@@ -1,0 +1,50 @@
+const express = require('express');
+const db = require("../../config/db");
+
+const router = express.Router();
+
+
+//컨트롤러 함수에서는 데이터베이스에서 부동산중개업소 정보를 조회하는 비즈니스 로직을 수행
+
+
+exports.getAgency = async(req,res) => {
+    const sgg_nm = req.query.sgg_nm;
+    const bjdong_nm = req.query.bjdong_nm;
+
+    console.log(`get agencies // sgg_nm: ${sgg_nm}, bjdong_nm: ${bjdong_nm}`);
+
+    try {
+        const query = `SELECT * FROM agencies WHERE sgg_nm = ? AND bjdong_nm = ?`;
+        const [rows, fields] = await db.query(query, [sgg_nm, bjdong_nm]);
+        console.log('agencies info ', rows);
+        res.json({ rows: rows });
+    } catch (err) {
+        console.error(err.stack)
+    }
+};
+
+// 중개업소 정보 조회 API
+exports.getOneAgency = async(req, res) => {
+  
+
+    const sgg_nm = req.query.sgg_nm;
+    const bjdong_nm = req.query.bjdong_nm;
+    const cmp_nm = '%'+req.query.cmp_nm+'%';
+
+    //console.log('searching an agency...........');
+    //console.log(`sgg_nm: ${sgg_nm}, bjdong_nm: ${bjdong_nm}, cmp_nm: ${cmp_nm}`);
+
+    try {
+        const query = `SELECT * FROM agencies WHERE sgg_nm = ? AND bjdong_nm = ? AND cmp_nm LIKE ?`;
+        //console.log(query);    
+            const [rows, fields]  = await db.query(query, [sgg_nm, bjdong_nm, cmp_nm]);          
+                //if (error) throw error;
+                 console.log('*****An agency info : ', rows);
+                 res.json({ rows: rows });
+        } catch (err) {
+        console.error(err.stack)
+    }
+    
+
+};
+
