+++ package.json
@@ -29,9 +29,10 @@
     "express-session": "^1.17.3",
     "jquery": "^3.6.4",
     "morgan": "^1.10.0",
-    "mysql": "^2.18.1",
+    "mysql2": "^3.3.0",
     "nodemon": "^2.0.22",
     "passport": "^0.6.0",
     "passport-local": "^1.0.0"
+
   }
 }
