+++ public/css/header.css
@@ -29,6 +29,12 @@ a{
     color:black;
 }
 
+a:hover{
+    color : black;
+    text-decoration: none;
+    
+}
+
 .topbar--container{
 
     display: flex;
