--- routers/index.js
@@ -1,13 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-router.get('/',(req,res)=>{
-
-    res.send('hello users');
-
-});
-
-
-
-
-module.exports = router;
